[level 1] Title: 콜라 문제, Time: 0.11 ms, Memory: 72.8 MB -BaekjoonHub

+++ 프로그래머스/1/132267. 콜라 문제/콜라 문제.java
@@ -2,9 +2,9 @@ class Solution {
     public int solution(int a, int b, int n) {
         int answer = 0;
         while(true){
-            long get = (long)(n/a) * b;
-            n = n%a + (int)get;
-            answer += (int)get;
+            int get = (n/a) * b;
+            n = n%a + get;
+            answer += get;
             if(n<a) break;
         } 
         return answer;
